hwdb: add accel sensor entry for Jumper EZpad Go
This commit fix the accelerometer orientation on the Jumper EZpad
Go tablet.

The tablet does not have its product name filled in dmi table, make
the match string a bit generic.  Here we assume that the use of a
KIOX000A + bios-vendor + chassis-type combo is unique enough to
match the currently available product in Jumper's x86 tablet series.

For future reference, as in 2019, the tablet has a dmialias of:

  dmi:bvnAmericanMegatrendsInc.:bvrZB-BI-11.6-SF133AR200-059-J \
  :bd05/21/2019:svnjumper:pnEZpad:pvrTobefilledbyO.E.M.:rvnTob \
  efilledbyO.E.M.:rnTobefilledbyO.E.M.:rvrTobefilledbyO.E.M.:c \
  vnTobefilledbyO.E.M.:ct31:cvrTobefilledbyO.E.M.: