tests: add test to cover OnFailure dependency removal
There are cases where it is desirable to be able to remove an
OnFailure dependency via a drop-in. These tests cover the changes
to support this:

Add a test which causes the OnFailure= dependency itself to fail
and ensure that a recursive dependency is not added even if the
dependency was cleared out via OnFailure= in a drop-in.

Signed-off-by: Peter Morrow <pemorrow@linux.microsoft.com>