core: allow OnFailure dependencies to be removed by fragments
There are cases where it's useful to be able to clear out an OnFailure
dependency that was installed somewhere higher up the fragment parse order.
Specifically if a user wants to set up a per unit type drop-in with an
OnFailure= dependency then we need a way clear out this dependency for
the original dependency otherwise we end up with a recursive dependency.
For example:

failure-handler@.service:

[Unit]
Description=Failure handler for service "%I"

[Service]
Type=oneshot
ExecStart=/bin/false

/etc/systemd/system/service.d/20-all-services-on-failure.conf:

[Unit]
OnFailure=failure-handler@%N.service

/etc/systemd/system/failure-handler@.service.d/failure-handler.conf:

[Unit]
OnFailure=

Signed-off-by: Peter Morrow <pemorrow@linux.microsoft.com>