core: add support for removing OnFailure dependencies
With per unit type drop-in files now supported (/etc/systemd/system/service.d,
/etc/systemd/system/mount.d etc) now available there is a use case for being
able to clear OnFailure dependencies via a drop in file. Using the
following set of units as an example:

/etc/systemd/system/failure-handler@.service:

[Unit]
Description=Failure handler for service "%I"

[Service]
Type=oneshot
ExecStart=/bin/myfailure-handler %i

/etc/systemd/system/service.d/20-all-services-on-failure.conf:

[Unit]
OnFailure=failure-handler@%N.service

/etc/systemd/system/failure-handler@.service.d/failure-handler.conf:

[Unit]
OnFailure=

Without the OnFailure= stanza we would end up creating a dependency
on failure-handler@failure-handler@failing-service.service (and so on in
a recursive fashion) which is not desirable in some cases.

This patch allows removal of the OnFailure dependency to break the
recursion.

Signed-off-by: Peter Morrow <pemorrow@linux.microsoft.com>